Fix: MySQL doesn't like long open connections.
This fix consists of two parts. First we extract only the data that we
need from the Django model, this is not stictly needed for this fix, but
why keeping Django objects in memory if we only need the username field.

Second part is to close the database connection after every user query.
Since this is the only Django database query we need, and the user(name)
is cached in the Client class (_user) we close the database connection.
(Django ORM will automaticly make a new connection when needed.)