Fix: make loading the playlist faster (tags backend).
The playlist was doing a lot of queries to retrieve the album and artist
information (in the template). With preloading this data, there are no
queries per track anymore, but just a single for the hole playlist.
(This affects only the 'tags backend'. The 'file backend' is still
utterly slow, because it reads all files to receive id3 information.
The relevant information for the playlist may be stored in the playlist
track objects. But this is for another time...