Use Django's UUIDField, to store musicbrainz uuid's
The UUIDField in fields.py was made because Django 1.6 didn't have an
UUIDField.. The version in fields.py was also broken because it was 32
characters, and the uuid stored in it, was with the dash-signs. So those
should have been stripped, or the field should have been 36 characters
long.

Another thing.. disable the artist UUIDs completely. There are some
things that need te be solved/implemented for songs with multiple
artists. A collaboration album can have multiple UUIDs for the
albumartist. Also a song can have a list of UUIDs for the artists. But
I'm not sure this list is the same sort order as the list of
'pronounceable'-artists. So this needs some work to properly fill the
artist UUID. I don't like the idea of calling the musicbrainz API.. But
I don't know if we have all the information localy in the music tags.